Expose logger and rename RunBackgroundWorkItemAsync
The `ILogger` on `BackgroundTaskService` is now `protected` instead of `private` so that it can be used in derived classes.

Breaking: `RunBackgroundWorkItemAsync` has been renamed to `InvokeWorkItemAsync` to avoid confusion with the non-abstract `RunBackgroundWorkItem` method. Additional documentation has been added to clarify usage.

In `InvokeWorkItemAsync`, `Activity.Current` is now always captured at the start of the invocation and restored at the end. Previously, this was only done if the `BackgroundWorkItem.ParentActivity` property was not null. This change ensures that background tasks that do not capture their parent activity are run as top-level activities.