Add BackgroundWorkItem.Cancel
`BackgroundWorkItem` includes a new `Cancel` method that can be used to explicitly request cancellation. If a work item is cancelled in this manner before it reaches the front of the queue it will not be invoked.

To enable explicit cancellation, `BackgroundWorkItem` is no longer a readonly struct so that the disposable pattern can be correctly implemented to dispose of the work item's private `CancellationTokenSource` on disposal.