Explicit work item cancellation (#7)
* Add BackgroundWorkItem.Cancel

`BackgroundWorkItem` includes a new `Cancel` method that can be used to explicitly request cancellation. If a work item is cancelled in this manner before it reaches the front of the queue it will not be invoked.

To enable explicit cancellation, `BackgroundWorkItem` is no longer a readonly struct so that the disposable pattern can be correctly implemented to dispose of the work item's private `CancellationTokenSource` on disposal.

* Allow explicit cancellation of work items

Note that, if a work item is explicitly cancelled before it reaches the front of the queue, `BackgroundTaskService` will not invoke the item's delegate.

* Add before/after work item processed events

* AppContext switch for running cancelled items

Changes the default behaviour so that work items that are already cancelled when they get to the front of the queue are no longer invoked.

This behaviour can be reversed by setting the `IntelligentPlant.BackgroundTasks.BackgroundTaskService.InvokeCancelledWorkItems` flag in `AppContext` to `true`.